Refactored unbox logic into sperate file
